raflow: DONE moving struct definitions from ws to rlib, bizlogic
